An updated version with multi-temperature MBPs and some basic input/output handling (still needsd to be tidied and tested)

+++ pack.hh
@@ -8,24 +8,24 @@ void packinit();
 int packsize();
 double * packbuffer();
 
-void pack(short num);
-void pack(long num);
-void pack(vector <long> &vec);
+void pack(int num);
+void pack(double num);
+void pack(vector <int> &vec);
 void pack(vector <double> &vec);
-void pack(vector< vector <long> > &vec);
+void pack(vector< vector <int> > &vec);
 void pack(vector< vector <double> > &vec);
 void pack(vector <string> &vec);
-void pack(vector< vector <FEV> > &vec, short fedivmin, short fedivmax);
-void pack(vector <HOUSE> &vec, long min, long max);
+void pack(vector< vector <FEV> > &vec, int fedivmin, int fedivmax);
+void pack(vector <HOUSE> &vec, int min, int max);
 
-void unpack(short &num);
-void unpack(long &num);
-void unpack(vector <long> &vec);
+void unpack(int &num);
+void unpack(double &num);
+void unpack(vector <int> &vec);
 void unpack(vector <double> &vec);
-void unpack(vector< vector <long> > &vec);
+void unpack(vector< vector <int> > &vec);
 void unpack(vector< vector <double> > &vec);
 void unpack(vector <string> &vec);
-void unpack(vector< vector <FEV> > &vec, short fedivmin, short fedivmax);
-void unpack(vector <HOUSE> &vec, long min, long max);
+void unpack(vector< vector <FEV> > &vec, int fedivmin, int fedivmax);
+void unpack(vector <HOUSE> &vec, int min, int max);
